Rename src/cmd_constants.h -> src/atm_cmd_constants.h
